made blocks into a react component
