+++ uccw/src/app/page.tsx
@@ -1,137 +1,19 @@
 'use client';
-import { useEffect } from 'react';
-import Matter, { World } from 'matter-js';
 import Image from 'next/image';
+import Blocks from "@/src/components/page/Blocks";
 
 export default function LandingPage() {
-    var start:boolean = false; //in the future, this should be made a state variable most likely
-                                
-
-    useEffect(() => {
-        
-
-        const height:number = window.innerHeight;
-        const width:number = window.innerWidth;
-
-        console.log("x: "+ width + " y: " + height);
-
-
-        // module aliases
-            const Engine = Matter.Engine,
-                Events = Matter.Events,
-                Render = Matter.Render,
-                Runner = Matter.Runner,
-                Body = Matter.Body,
-                Bodies = Matter.Bodies,
-                Composite = Matter.Composite;
-
-            // create an engine
-            const engine = Engine.create();
-            engine.gravity.scale = .00005;
-
-            // create a renderer
-            const render = Render.create({
-                element: document.body,  //creates canvas element to render to
-                engine: engine,
-                options: {
-                    width,
-                    height,
-                    background: 'transparent',
-                    wireframes: false,
-                }
-            });
-            render.canvas.style.position = 'fixed';
-            render.canvas.style.top = '0';
-            render.canvas.style.left = '0';
-            render.canvas.style.zIndex = '-5';
-
-
-            function spawn() {
-                const randX = Math.random() * innerWidth;
-                const randSize = 20 + Math.random() * 80;
-                const randGrav = Math.random() / 50;
-                const box = Bodies.rectangle(randX, -100, randSize, randSize, {
-                    frictionAir: randGrav,
-                    render: {
-                        sprite: {
-                            texture: '/square.png',
-                            xScale: randSize/ 445,
-                            yScale: randSize / 446,
-                        } 
-                    }});
-                const spin = (Math.random() - 0.5) * 0.4; //goes too fast sometimes but its funny
-                Body.setAngularVelocity(box, spin);
-                Composite.add(engine.world, box);
-            }
-
-            var last = Date.now();
-            
-            window.addEventListener('keydown', (event) => {
-                if (event.key === " ") {
-                    console.log("Space")
-                    event.preventDefault();
-                    start = !start;
-                }
-            });
-
-            Events.on(engine, 'beforeUpdate', () => {
-                if (!start) return;
-                const now = Date.now();
-                if (now - last > 1000) {
-                    spawn();
-                    last = now;
-                }
-            })
-            
-            // run the renderer
-            Render.run(render);
-            // create runner
-            const runner = Runner.create();
-            // run the engine
-            Runner.run(runner, engine);
-
-
-            //maybe use setposition to bring back to top? itd be less random.
-            Events.on(engine, 'afterUpdate', () => {
-                const bodies = Composite.allBodies(engine.world);
-                for (const body of bodies) {
-                    const {x, y} = body.position;
-                    if (y > height + 1000) {
-                        World.remove(engine.world, body);
-                        console.log("DELETED");
-                    }}
-            });
-
-            window.addEventListener("resize", (handleResize))
-            function handleResize() {
-                render.canvas.width = window.innerWidth;
-                render.canvas.height = window.innerHeight;
-                render.options.height = window.innerHeight;
-                render.options.width = window.innerWidth;
-                Render.lookAt(render, {
-                    min: {x: 0, y: 0},
-                    max: {x: innerWidth, y: innerHeight},
-                })
-            }
     
-    return () => {
-        
-        Render.stop(render);
-        Runner.stop(runner);
-        render.canvas.remove();
-        
-    };
-
-
-    });
 
     return (
-        <div className="select-none">
+        
+        <div className="select-none overflow-hidden">
+            <Blocks/>
             <div className="flex items-center justify-around px-10 py-30">
-                <div className="max-w-2xl z-10">
-                    <h1 className='font-extrabold text-5xl text-white text-balance leading-relaxed drop-shadow-lg/50 select-none'>University of Nebraska Lincoln Coding Club (WIP)</h1>
+                <div className="max-w-2xl z-10 ">
+                    <h1 className='font-extrabold text-5xl text-white text-balance leading-relaxed drop-shadow-lg/50 select-none'>University of Nebraska Lincoln Coding Club</h1>
                 </div>
-                <div className="drop-shadow-2xl/25 drop-shadow-white -z-10">
+                <div className="drop-shadow-2xl/25 drop-shadow-white z-10">
                     <Image 
                     src='/landinglogotest.png' 
                     alt='Logo'
@@ -142,9 +24,9 @@ export default function LandingPage() {
             </div>
             {/*top block*/}
             <div className="absolute w-screen h-1/8 top-0 left-0 -z-1 bg-bg-primary"></div>
-            {!start &&<div className="fixed flex bottom-5 left-1/2 transform -translate-x-1/2  items-center text-xl text-white">
-                <p>press space to start</p>
-            </div>}
+
+            
+            
         </div>
     )
